[memo] Cached values are reused without removal/reinsertion.
Moves nearly all of the free function's code into methods on MemoStore.

Pulls memo free functions to crate root.

MemoStorage becomes responsible for argument comparison.

Moves MemoStore/MemoStorage to their own module.

MemoStore becomes generic over the location it accepts.
(defaults to topo::Id still).
